Split up x11rb::connection
This commit splits the cookies and ExtensionInformation to separate
files to shrink x11rb::connection a bit.
